% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{collect.tbl_GeopkgConnection}
\alias{collect.tbl_GeopkgConnection}
\title{Retrieves data into a local tibble}
\usage{
\method{collect}{tbl_GeopkgConnection}(x, ..., n = Inf, warn_incomplete = TRUE, cte = FALSE)
}
\arguments{
\item{x}{A lazy data frame backed by a database query.}

\item{...}{other parameters passed to methods}

\item{n}{number of rows to fetch, default all}

\item{warn_incomplete}{Warn if n is less than the number of result rows?}

\item{cte}{use common table expressions}
}
\value{
a local sf tibble
}
\description{
Retrieves data into a local tibble
}
