% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{connectGPKG}
\alias{connectGPKG}
\title{create connection to a geopackage database file}
\usage{
connectGPKG(gpkgname)
}
\arguments{
\item{gpkgname}{the filename, passed to DBI::dbConnect}
}
\value{
a connection object including structures required to create sf objects
}
\description{
create connection to a geopackage database file
}
\examples{
library(dplyr)
ncf <- system.file("gpkg/nc.gpkg", package="sf")
nc <- connectGPKG(ncf)
nc
tbl(nc, "nc.gpkg")
nicenames <- tibble::tibble(NAME=c("Northampton", "Stokes"))
f <- tbl(nc, "nc.gpkg")
collect(semi_join(f, nicenames, copy=TRUE))

}
