% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\docType{class}
\name{GeopkgConnection-class}
\alias{GeopkgConnection-class}
\title{Connection class for a geopkg flie}
\description{
Connection class for a geopkg flie
}
\section{Slots}{

\describe{
\item{\code{CRS}}{tbl.}
}}

