% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{show,GeopkgConnection-method}
\alias{show,GeopkgConnection-method}
\title{show method for geopackage connection object}
\usage{
\S4method{show}{GeopkgConnection}(object)
}
\arguments{
\item{object}{the connection object}
}
\value{
prints a summary of the object
}
\description{
show method for geopackage connection object
}
