% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{tbl.GeopkgConnection}
\alias{tbl.GeopkgConnection}
\title{Create table from geopackage src}
\usage{
\method{tbl}{GeopkgConnection}(src, from, ...)
}
\arguments{
\item{src}{the geopkg connection}

\item{from}{table name}

\item{...}{passed to methods}
}
\description{
Create table from geopackage src
}
